Skip lines that might result in match-all pattern
Before this change the ignore string got appended with the line as-it-is
and if there were a rule[1] that accidentally got defined with a `*,` as
the prefix, it would get appended to the list and becomes a match-all
rule, effectively breaking plugins that rely on this functionality. One
such plugin being `Shougo/neosnippet.vim`.

Since it's hard to understand the intent of the rule and add necessary
substitutions, it's better to just skip the line.

1: https://github.com/getsentry/onpremise/pull/193